Support additional Postgres types
Add supports for:
- json
- jsonb
- time
- timetz
- timestamp
- timestamptz
- bytea

To support JSON implemented a somewhat pluggable postgres-to-typescript type
mapping via TypeAllocator which is able to emit type aliases and imports in the
generated declaration file.

Also:

- Moved some code out of index.ts to generator.ts in CLI package to make
it easier to test and provide better segregation
- Added an option to TypeAllocator to just emit unmapped types with
their Postgres name. This is not yet exposed through CLI but is useful
for fixing up types (e.g. enums or composite types) that are not yet
supported.
- Added some example code that makes use of the JSON support

Signed-off-by: Silas Davis <silas@monax.io>