chore: remove unused locals and parameters
This enables noUnusedLocals and noUnusedParameters TypeScript compiler
options to check for unused parameters, imports, and variables.
Unfortunately the antlr4ts generator creates code that violates this, so
it's necessary to make a trade-off and pick the lesser evil. Given how
many other places in the code-base had dead code, I think it's more
valuable to have it enabled and deal with manually editing the antlr4ts
output. There's an open issue about it
https://github.com/tunnelvisionlabs/antlr4ts/issues/412 so it might
eventually be fixed upstream.