Add support for empty array spread in WHERE IN
 - Extend `replaceIntervals` to handle empty where in (affects both SQL
   and TS)
 - Update docs & tests

**Implementation**

Function `replaceIntervals` can now detect, if the interval is part of
WHERE IN condition and can replace larger part of the query with an
tautological condition. The solution should work safetly for queries
that use `()` intentionally somewhere (comments) and that have the WHERE
IN nested in other conditional chains. If the replacement occurs, SQL
comment is added.

**Motivation**

Support for empty where-in is way more vital use-case
compared to support for values. Values are used most of the time for
creates and it does not make sense to have a create variant query for
the results: this should be validated beforehand instead.

If you do have a reasonably large query that has 3 optional WHERE IN
filters that you can construct with a query builder, with PgTyped, you'd
need to create all the possible combinations for the slight variances.

Fixes: https://github.com/adelsz/pgtyped/issues/273

Related: https://github.com/adelsz/pgtyped/issues/221
Related: https://github.com/adelsz/pgtyped/issues/314
Related: https://github.com/adelsz/pgtyped/issues/156